% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{clanElectrons}
\alias{clanElectrons}
\title{clanElectrons}
\usage{
clanElectrons()
}
\description{
Package for computation of electronic stopping power and density effect (Sternheimer)
This is a simple dummy function to ease access to the help index.

 clanElectrons()

 ?clanElectrons # gives you an index of functions in package

 Status (August 28, 2022):
 1. We miss handling of delta for low energy electrons (delta = 0).

 2. Further simplifications could be good.

Main functions for computation of the density-effect correction are:

   dat <- Sternheimer.delta.param(MeV, dat)

   dat <- Sternheimer.delta.exact(MeV, dat)

Note that dat is a list with input and output parameters. The .parm version
is just the simple use of the Sternheimer fitting parameters (see e.g. the 1984
paper with parameters for many materials). The .exact version is the complex
model based on oscillators (binding electies and subshell atomic structure) and
dielectric (optical) stopping.

 Main functions for stopping power power computations are:
    electronic.MSP(MeV, dat, delta)
    electronic.MSP.restricted(MeV,delta.keV, dat, delta)

Main validation functions (comparison against ICRU-90 data):
  demo.Sternheimer.graphite()
  demo.Sternheimer.water()
}
