% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{delta.Sternheimer}
\alias{delta.Sternheimer}
\title{delta.Sternheimer
Not so easy as Sternheimer}
\usage{
delta.Sternheimer(
  MeV = 1,
  I = 81,
  Z = 6,
  A = 12.011,
  rho = 2.265,
  Sternheimer.tab.id = "graphite-1",
  C.model = "fixed",
  df.Sternheimer = NULL,
  verbose = FALSE
)
}
\description{
delta.Sternheimer
Not so easy as Sternheimer
}
