% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{demo.electronic.MSP.restricted}
\alias{demo.electronic.MSP.restricted}
\title{demo.electronic.MSP.restricted}
\usage{
demo.electronic.MSP.restricted()
}
\description{
Demonstration of how to use some the restricted electronic stopping
power. Here we look at water with cutoff values equal to 1, 10, 100, and 1000 keV.
}
\details{
None
}
