% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{demo.Sternheimer.delta.param}
\alias{demo.Sternheimer.delta.param}
\title{demo.Sternheimer.delta.param}
\usage{
demo.Sternheimer.delta.param()
}
\description{
Demonstration of how to useSternheimer.delta.param().
}
\details{
Notes:
none
}
