% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{demo.Sternheimer.delta.plot}
\alias{demo.Sternheimer.delta.plot}
\title{demo.Sternheimer.delta.plot
Delta vs. energy for two different models (Al: En=0 or not)}
\usage{
demo.Sternheimer.delta.plot()
}
\description{
demo.Sternheimer.delta.plot
Delta vs. energy for two different models (Al: En=0 or not)
}
