% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{demo.Sternheimer.graphite}
\alias{demo.Sternheimer.graphite}
\title{demo.Sternheimer.graphite
Computation of electronic stopping power and density effect for graphite
using Sternheimer model as described in ICRU-90}
\usage{
demo.Sternheimer.graphite()
}
\description{
demo.Sternheimer.graphite
Computation of electronic stopping power and density effect for graphite
using Sternheimer model as described in ICRU-90
}
