% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{demo.Sternheimer.water}
\alias{demo.Sternheimer.water}
\title{demo.Sternheimer.water
Computation of electronic stopping power and density effect for liqud water
using Sternheimer model as described in ICRU-90}
\usage{
demo.Sternheimer.water()
}
\description{
demo.Sternheimer.water
Computation of electronic stopping power and density effect for liqud water
using Sternheimer model as described in ICRU-90
}
