% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{demo.stopping.power.computations}
\alias{demo.stopping.power.computations}
\title{demo.stopping.power.computations
Not so easy as Sternheimer}
\usage{
demo.stopping.power.computations()
}
\description{
demo.stopping.power.computations
Not so easy as Sternheimer
}
