% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{demo.stopping.power.for.water}
\alias{demo.stopping.power.for.water}
\title{demo.stopping.power.for.water
Not so easy as Sternheimer}
\usage{
demo.stopping.power.for.water()
}
\description{
demo.stopping.power.for.water
Not so easy as Sternheimer
}
