% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{demo.stopping.power.plot}
\alias{demo.stopping.power.plot}
\title{demo.stopping.power.plot
Not so easy as Sternheimer}
\usage{
demo.stopping.power.plot()
}
\description{
demo.stopping.power.plot
Not so easy as Sternheimer
}
