% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST-estar.R
\docType{data}
\name{df.NIST.xray.materials}
\alias{df.NIST.xray.materials}
\title{df.NIST.xray.materials}
\format{
A data frame with 8240 rows and 15 variables:
\describe{
  \item{Z}{Element number (=0 for composite materials))}
  \item{symbol}{Element symbols like "H", "He", "Br".}
  \item{id.xray}{Name of element or material from NIST}
  \item{Z.A.ratio}{Z/A}
  \item{I.xray}{Mean excitation energy (eV)}
  \item{rho.xray}{Density (in g/cm3).}
}
}
\source{
Generated for package examples
}
\usage{
df.NIST.xray.materials
}
\description{
NIST material data from X-Ray Mass Attenuation Coefficients database (not Estar)
# From https://physics.nist.gov/PhysRefData/XrayMassCoef/tab1.html
X-Ray Mass Attenuation Coefficients
Table 1. Material constants assumed in the present evaluations for elemental media.
Values are given for the ratio of atomic number-to-mass Z/A, the mean excitation energy I, and the density ρ.
Some density values are only nominal; those for Z = 85 and 87 were arbitrarily set to 10 in order to complete the calculations.
Variable 1: Z
Variable 2: Element
Variable 3: Name
Variable 4:  Z/A = Z.A.ratio
Variable 5: I	in eV
Variable 6: Density in g/cm3
Downloaded: February 7, 2026 by Claus E. Andersen
}
\keyword{datasets}
