% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{electronic.MSP.Bethe}
\alias{electronic.MSP.Bethe}
\title{electronic.MSP.Bethe
Not so easy as Sternheimer}
\usage{
electronic.MSP.Bethe(MeV = 1, dat = NULL(), delta = NA)
}
\description{
electronic.MSP.Bethe
Not so easy as Sternheimer
}
