% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{electronic.MSP}
\alias{electronic.MSP}
\title{electronic.MSP}
\usage{
electronic.MSP(MeV = 1, dat = NULL(), delta = 0)
}
\arguments{
\item{MeV}{kinetic energy of the electron in MeV}

\item{dat}{list with parameters}

\item{delta}{delta parameter}
}
\value{
mass stopping power
}
\description{
Computation of mass electronic stopping power (MSP) according to
ICRU-90 eq. 4.8 (page 22). The correction for the density effect (delta)
has to be supplied by the user in a separate argument.

dat$Z = atomic number
dat$A = atomic mass
dat$I = mean excitation energy in eV
delta = density-effect correction
}
\details{
Notes:
The computed MSP is in units of MeV pr. g/cm2.
}
