% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST-estar.R
\name{get.estar.from.id}
\alias{get.estar.from.id}
\title{get.estar.from.id}
\usage{
get.estar.from.id(
  MeV = c(0.1, 1, 10),
  what = "MSP.el",
  id = "water.liquid",
  data = df.NIST.estar.SP,
  rule = 1
)
}
\arguments{
\item{MeV}{= vector of kinetic energies of the electron in MeV}

\item{what}{="MSP.el", MSP.rad", "MSP.tot", "range.CSDA", "rad.yield", "density.effect", "I", "rho", "Z.A.ratio"}

\item{id}{= "hydrogen", "helium" .. "californium", "water.liquid", "air.dry", "pmma", "alanine", "carbon"}

\item{data}{= date frame with data (df.NIST.estar.SP)}

\item{rule}{= rule for approx: 1=no extrapolation, 2=use nearest value for extrapolation}
}
\description{
Extract data from the df.NIST.estar.SP data for given id (=name of element or material).
df.NIST.estar.SP is included in the package.
}
\details{
Notes:
The  MSP's are in units of MeV pr. g/cm2.
The I is in eV
The density (rho) is in g/cm3
}
