% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST-estar.R
\name{get.estar.from.Z}
\alias{get.estar.from.Z}
\title{get.estar.from.Z}
\usage{
get.estar.from.Z(
  MeV = c(0.1, 1, 10),
  what = "MSP.el",
  Z = 1,
  data = df.NIST.estar.SP,
  rule = 1
)
}
\arguments{
\item{MeV}{= vector of kinetic energies of the electron in MeV}

\item{what}{="MSP.el", MSP.rad", "MSP.tot", "CSDA", "rad.yield" or "density.effect"}

\item{Z}{= element number (1,2,3 .. 98) (composite materials have Z=0)}

\item{data}{= date frame with data (df.NIST.estar.SP)}

\item{rule}{= rule for approx: 1=no extrapolation, 2=use nearest value for extrapolation}
}
\description{
Extract data from the df.NIST.estar.SP data for given Z.
df.NIST.estar.SP is included in the package.
}
\details{
Notes:
The  MSP's are in units of MeV pr. g/cm2.
The I is in eV
The density (rho) is in g/cm3
}
