% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST-estar.R
\name{get.NIST.estar}
\alias{get.NIST.estar}
\title{get.NIST.estar}
\usage{
get.NIST.estar(
  MeV = c(0.1, 1, 10),
  what = "MSP.el",
  id = "water.liquid",
  data = df.NIST.estar.SP
)
}
\arguments{
\item{MeV}{= vector of kinetic energies of the electron in MeV}

\item{what}{="MSP.el", MSP.rad", "MSP.tot", "CSDA", "rad.yield" or "density.effect"}

\item{id}{= "water.liquid", "air.dry", "pmma", "alanine"}

\item{data}{= date frame with data (df.NIST.estar.SP)}
}
\description{
Extract data from the df.NIST.estar.SP data frame included
with the package.
}
\details{
Notes:
The  MSP's are in units of MeV pr. g/cm2.
The I is in eV
The density (rho) is in g/cm3
}
