% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{read.Sternheimer.data}
\alias{read.Sternheimer.data}
\title{read.Sternheimer.data
Not so easy as Sternheimer}
\usage{
read.Sternheimer.data(print.wanted = FALSE)
}
\description{
read.Sternheimer.data
Not so easy as Sternheimer
}
