% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{restricted.electronic.MSP.Bethe}
\alias{restricted.electronic.MSP.Bethe}
\title{restricted.electronic.MSP.Bethe
Not so easy as Sternheimer}
\usage{
restricted.electronic.MSP.Bethe(
  MeV = 1,
  delta.keV = 10,
  dat = NULL,
  delta = NA
)
}
\description{
restricted.electronic.MSP.Bethe
Not so easy as Sternheimer
}
