% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{sensitivity.electronic.MSP.Bethe.dlog.dlogI}
\alias{sensitivity.electronic.MSP.Bethe.dlog.dlogI}
\title{sensitivity.electronic.MSP.Bethe.dlog.dlogI
Not so easy as Sternheimer}
\usage{
sensitivity.electronic.MSP.Bethe.dlog.dlogI(
  MeV = 1,
  I = 81,
  Z = 6,
  A = 12.011,
  rho = 2.265,
  C.model = "fixed",
  df.Sternheimer = NULL
)
}
\description{
sensitivity.electronic.MSP.Bethe.dlog.dlogI
Not so easy as Sternheimer
}
