% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{Sternheimer.beta.threshold}
\alias{Sternheimer.beta.threshold}
\title{Sternheimer.beta.threshold}
\usage{
Sternheimer.beta.threshold(L, dat)
}
\description{
Helper function computing the threshold below which an insulator will have
zero density-effect correction. Formulated both as a beta.threshold and MeV.threshold (i.e.
kinitic energy expressed in MeV).
Function returns an updated version of dat.
}
