% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{Sternheimer.data.read}
\alias{Sternheimer.data.read}
\title{Sternheimer.data.read}
\usage{
Sternheimer.data.read(print.wanted = FALSE)
}
\arguments{
\item{print.wanted}{= TRUE or FALSE}
}
\description{
Data from Sternheimers approximitive fit to the density-effect correction
in the paper:
R.M. Sternheimer (Brookhaven), M.J. Berger (NBS) and S.M. Seltzer (NBS): Density effect
for the ionization loss of charged particles in various substances. Atomic Data and Nuclear
Data Tables 30,26 l-27 1 (1984).
}
\details{
Notes:
function returns a data frame with fitted parameters for a few selected materials.
}
