% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{Sternheimer.delta.exact}
\alias{Sternheimer.delta.exact}
\title{Sternheimer.delta.exact}
\usage{
Sternheimer.delta.exact(MeV = 1, dat = NULL)
}
\arguments{
\item{dat}{list with parameters}

\item{dat$MeV}{= electron kinetic energy}

\item{dat$nlev}{= number of sub-shells}

\item{dat$Z}{= atomic number}

\item{dat$A}{= atomic mass}

\item{dat$rho.density}{= density in g/cm3}

\item{dat$fvec}{= sub-shell occupancy level (used in computation)}

\item{dat$Evec}{= sub-shell binding energy (used in computation)}

\item{dat$fvec.org}{= sub-shell occupancy level (NOT used in computation)}

\item{dat$Evec.org}{= sub-shell binding energy (NOT used in computation)}

\item{dat$I}{= mean excitation energy in eV}

\item{dat$plot.wanted}{= TRUE or FALSE,}
}
\description{
Computation of exact Sternheimer density correction.

By exact, we just mean that this is not the simple
parameter fitting solution. The ICRU-90 terminology
(p. 26) is followed as close as possible).
}
\details{
Notes:
(1) The equations for mu.st and L are solved using a search strategy
    (coarse + fine + regression) which may fail, for example, if the first interval
    dose not include the root. Graphs can ne plotted to inspect  the procedure.
(2) MeV is the kinetic energy and only one value can be computeted (not vectorized).
(3) Only computations for single atoms has been implemented (not multi-atom materials).
(4) The data fit well the ICRU-37 delta values for high energies (e.g. at 100 MeV), but
    less well for lower energies (e.g. at 1 MeV). The source of deviation has not been
    identified.
(5) The main use of this work is to try to gain some insights into the procedure.
    Like how conductors are handled. And that insulators have delta = 0 for energies
    below a certain threshold beta0 (not roots).
(6) Why is this insight relevant? This is needed as the text is not super clear and
    since the  whole concept of dialetric / optical modelling is new to me (e.g. the
    use of plasma energy).

Main references:
(1) ICRU-90: Key data for ionizing-radiation dosimetry: Measurement standards
    and applications (2014/2016).
(2) DENSITY EFFECT FOR THE IONIZATION LOSS OF CHARGED PARTICLES
    IN VARIOUS SUBSTANCES
    by R. M. STERNHEIMER (Brookhaven), M. J. BERGER (NBS) and S. M. SELTZER (NBS).
    ATOMIC DATA AND NUCLEAR DATA TABLES 30,26 l-27 1 ( 1984)
(3) G4DensityEffectCalculator.cc by Matthew Strait <straitm@umn.edu> 2019
}
