% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{Sternheimer.delta.param}
\alias{Sternheimer.delta.param}
\title{Sternheimer.delta.param}
\usage{
Sternheimer.delta.param(MeV = 1, dat = NA)
}
\description{
Sternheimer.delta.param
}
