% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{Sternheimer.f.root.L}
\alias{Sternheimer.f.root.L}
\title{Sternheimer.f.root.L}
\usage{
Sternheimer.f.root.L(L, dat)
}
\description{
Helper function (L) for exact Sternheimer density correction
# Equation suitable for finding L (ell) (eq. 4.28 in ICRU-90). We search a value of L such that
Sternheimer.f.root.L(L, dat) is approximately zero.
}
