% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{Sternheimer.set.to.conductor}
\alias{Sternheimer.set.to.conductor}
\title{Sternheimer.set.to.conductor}
\usage{
Sternheimer.set.to.conductor(dat)
}
\description{
Ascertain that material is treated as conductor in Sternheimer delta computations

 dat <- Sternheimer.set.to.conductor(dat)

The binding energy of the outer subshell is set to zero
}
