% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{Sternheimer.set.to.insulator}
\alias{Sternheimer.set.to.insulator}
\title{Sternheimer.set.to.insulation}
\usage{
Sternheimer.set.to.insulator(dat)
}
\description{
Ascertain that material is treated as insulator in Sternheimer delta computations

 dat <- Sternheimer.set.to.insulation(dat)

The binding energy of the outer subshell is NOT set to zero
}
