% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{clanElectrons}
\alias{clanElectrons}
\title{clanElectrons}
\usage{
clanElectrons()
}
\description{
Package for computation of electronic stopping power and density effect (Sternheimer)
Dummy function to ease acces to the help index.

 clanElectrons()

 ?clanElectrons # gives you an index of functions in package

 #' Status (August 10, 2022):

 1. Bethe formula and Sternheimer.exact.delta agres perfectly with ICRU-90
 for water, but NOT for graphite.

 2. More work needed for the lower energies for insulators and conductors.
 Clean up of examples are needed.

 3. Consider if material data could be stored.

 4. Perhaps we should make a set of functions:

  demo.MSP.water

  demo.MSP.graphite

  demo.alanine

  demo.MSP.Al

  demo.MSP.Be

  etc.
}
