% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{demo.Bragg.rule.test}
\alias{demo.Bragg.rule.test}
\title{demo.Bragg.rule.test}
\usage{
demo.Bragg.rule.test()
}
\description{
This function test the application of Bragg's rule
where, for example, the stopping power for water is computed pased on the
stopping powers of H and O.

What we can learn from this function:

(1) For a compound like water, it is better to use effective values for Z and A than Braggs rule.
    So, for water it is best to apply Z = 10 and A = 18.0158 in a single call to the Bethe formula.
    The Bragg rule by taking the Sel/rho(H2O)  = w.H * Sel/rho(H) + w.O * Sel/rho(O) can easily be
    a some percent off.

(2) The remaining deviations between the Bethe estimate using Z=2x1+8=10 and A=18.0158 originates from
    the delta value used for the density effect.
}
\details{
None
}
