% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{demo.graphite}
\alias{demo.graphite}
\title{demo.graphite}
\usage{
demo.graphite(MeV = 1)
}
\description{
Computation of electronic stopping power and density effect for graphite
}
\details{
#' We use the "exact" Sternheimer model as described in ICRU-90.
Graphite is treated as a conductor (nc=1). We compare against ICRU-90 computations.
}
