% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{demo.Sternheimer.delta.exact.plot}
\alias{demo.Sternheimer.delta.exact.plot}
\title{demo.Sternheimer.delta.exact.plot}
\usage{
demo.Sternheimer.delta.exact.plot()
}
\description{
Delta vs. energy for two different models (Al: En=0 or not)
}
\details{
none.
}
