% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{demo.Sternheimer.delta.param}
\alias{demo.Sternheimer.delta.param}
\title{demo.stopping.power.validation
Not so easy as Sternheimer}
\usage{
demo.Sternheimer.delta.param()
}
\description{
demo.stopping.power.validation
Not so easy as Sternheimer
}
