% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST-estar.R
\docType{data}
\name{df.NIST.estar.SP}
\alias{df.NIST.estar.SP}
\title{df.NIST.estar.SP}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{MeV}{Date of the sale (Date)}
  \item{id}{Product name (character)}
  \item{Z}{Number of units sold (integer)}
}
}
\source{
Generated for package examples
}
\usage{
df.NIST.estar.SP
}
\description{
Extract data from the df.NIST.estar.SP data for given id (=name of element or material).
df.NIST.estar.SP is included in the package.
}
\keyword{datasets}
