% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{electronic.MSP.Bethe}
\alias{electronic.MSP.Bethe}
\title{electronic.MSP.Bethe
Not so easy as Sternheimer}
\usage{
electronic.MSP.Bethe(
  MeV = 1,
  I = 81,
  Z = 6,
  A = 12.011,
  rho = 2.265,
  Sternheimer.tab.id = "graphite-1",
  C.model = "fixed",
  delta.fixed = NA,
  df.Sternheimer = NULL,
  verbose = FALSE
)
}
\description{
electronic.MSP.Bethe
Not so easy as Sternheimer
}
