% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bethe-stopping-power-computations-001.R
\name{Sternheimer.delta.param}
\alias{Sternheimer.delta.param}
\title{Sternheimer.delta.param}
\usage{
Sternheimer.delta.param(MeV = 1, dat = NA)
}
\arguments{
\item{MeV}{= kinetic energy of the electron (in MeV)}

\item{dat}{= list with parameters}
}
\value{
a list (dat)
}
\description{
Compute the density-effect correction based on the fitting parameters
}
\details{
The fitted parameters can be found, for example, in the paper:
- R.M. Sternheimer (Brookhaven), M.J. Berger (NBS) and S.M. Seltzer (NBS): Density effect
for the ionization loss of charged particles in various substances. Atomic Data and Nuclear
Data Tables 30,26 l-27 1 (1984).

The function returns an approximate value for the density-effect correction (delta).
See also Andero et al. "Fundamentals of ionizing radiation dosimetry" (2017) p. 74.

The output ia returned as a list (dat):
  dat$param.MeV =  MeV
  dat$param.delta = delta

First create a material (i.e. make a dat-list) called dat or something like than.

Then call the function for the given energy MeV:

dat <- Sternheimer.delta.param(MeV, dat)

dat$param.C = Sternheimer parameter (see 1984 paper)
dat$param.X0 = Sternheimer parameter (see 1984 paper)
dat$param.X1 = Sternheimer parameter (see 1984 paper)
dat$param.a = Sternheimer parameter (see 1984 paper)
dat$param.m = Sternheimer parameter (see 1984 paper)
}
