% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sternheimer-exact-001.R
\name{Sternheimer.f.root.mu.st}
\alias{Sternheimer.f.root.mu.st}
\title{Sternheimer.f.root.mu.st}
\usage{
Sternheimer.f.root.mu.st(mu.st, dat)
}
\arguments{
\item{mu.st}{= mu.st-parameter}

\item{dat}{= list with parameters}
}
\description{
Helper function (mu.st) for exact Sternheimer density correction

# Equation suitable for finding mu.st (eq. 4.29 in ICRU-90). We search a value of mu.st such that
Sternheimer.f.root.mu.st(mu.st, dat) is approximately zero.
}
